feat: refactor to notificationService and finish show: 'notiMinSize = 10' feature
